% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decision-set-feasibility}
\alias{decision-set-feasibility}
\alias{feasible}
\alias{infeasible}
\alias{infeasible_reasons}
\alias{feasible,DecisionSet-method}
\alias{infeasible,DecisionSet-method}
\alias{infeasible_reasons,DecisionSet-method}
\title{Subset a decision set by feasibility}
\usage{
feasible(object)

\S4method{feasible}{DecisionSet}(object)

infeasible(object)

\S4method{infeasible}{DecisionSet}(object)

infeasible_reasons(object)

\S4method{infeasible_reasons}{DecisionSet}(object)
}
\arguments{
\item{object}{A \linkS4class{DecisionSet} object.}
}
\value{
\code{feasible()} and \code{infeasible()} return a \linkS4class{DecisionSet} object.
\code{infeasible_reasons()} returns a tibble with columns \code{scenario} and
\code{message}.
}
\description{
\code{feasible()} and \code{infeasible()} partition a \linkS4class{DecisionSet} into the
instances that did and did not reach a decision, each returning a
\code{DecisionSet} of the same shape. \code{infeasible_reasons()} returns the reasons
table for the instances that failed.
}
\seealso{
\code{\link[=decide]{decide()}}, \code{\link[=is_feasible]{is_feasible()}}
}
