% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/fitted-class.R
\name{sense}
\alias{sense}
\alias{is_minimization}
\alias{sense,OptimizationProblem-method}
\alias{is_minimization,OptimizationProblem-method}
\alias{sense,DecisionFocusedLasso-method}
\alias{is_minimization,DecisionFocusedLasso-method}
\title{Objective sense of a problem}
\usage{
sense(object)

is_minimization(object)

\S4method{sense}{OptimizationProblem}(object)

\S4method{is_minimization}{OptimizationProblem}(object)

\S4method{sense}{DecisionFocusedLasso}(object)

\S4method{is_minimization}{DecisionFocusedLasso}(object)
}
\arguments{
\item{object}{An \linkS4class{OptimizationProblem} object, such as one returned by
\code{\link[=knapsack_problem]{knapsack_problem()}} or \code{\link[=optimization_problem]{optimization_problem()}}.}
}
\value{
\code{sense()} returns a character scalar, \code{"min"} or \code{"max"}.
\code{is_minimization()} returns a single logical, \code{TRUE} when the sense is
\code{"min"}.
}
\description{
Reports whether an optimisation problem minimises or maximises. \code{sense()}
returns the direction as a string; \code{is_minimization()} returns a logical,
\code{TRUE} for \code{"min"}.
}
\examples{
problem <- knapsack_problem()
sense(problem)
is_minimization(problem)

}
\seealso{
\code{\link[=optimization_problem]{optimization_problem()}}

Other dflasso problems: 
\code{\link{OptimizationProblem-class}},
\code{\link{capital_allocation_problem}()},
\code{\link{knapsack_problem}()},
\code{\link{optimization_problem}()},
\code{\link{shortest_path_problem}()},
\code{\link{solve_decision}()}
}
\concept{dflasso problems}
